refactor(core): #238 observer 注入 TurnContext
audit 从全局函数改为 Observer 结构体:
- 新增 Observer(持有 storage_root)注入 TurnContext
- observe/mod.rs 合并原 audit.rs + observe/audit.rs
- engine.rs 审计调用改为 self.observer.audit_permission/audit_tool_execution
- 删除 crate::audit 模块(AuditEntry/append_audit_log 合入 observe)
- 插件管理操作(skill/mcp)的 audit 从 observe 模块 re-export

turn 级审计不再依赖全局 STORAGE_ROOT(从 Observer.storage_root 取);
非 turn 级(插件管理)暂保留全局 storage_root 依赖,待后续清理。
