+++ crates/tiangong-core/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod agent_config;
 pub mod agent_input;
 pub mod approval_store;
-pub mod audit;
 pub mod context;
 pub mod core;
 pub mod core_config;
